low-level

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,11 +5,11 @@
  * @n: size
  * Return: 1 or fib(n)
  */
-double fib(int n)
+long int fib(int n)
 {
-	double a = 1;
-	double b = 1;
-	double out = 0;
+	long int a = 1;
+	long int b = 1;
+	long int out = 0;
 	int i = 0;
 
 	for (; i < n; i++)
@@ -31,7 +31,7 @@ int main(void)
 	int i = 1;
 
 	for (; i < n; i++)
-		printf("%f, ", fib(i));
-	printf("%f\n", fib(n));
+		printf("%ld, ", fib(i));
+	printf("%ld\n", fib(n));
 	return (0);
 }
